+++ mrv2/docs/HISTORY.md
@@ -1,8 +1,11 @@
 v0.3.9
 ------
 - Added Search on Hotkeys for functions and hotkeys.
-- Updated all Python on every OS to 3.10.9.
+- Updated all Python on every OS to 3.10.9, which is the sanctioned Python
+  for VFX Platform 2023.
 - Fixed resizing of dock and close button on macOS.
+- Improved Docker building by not cloning the git repository in the Dockerfile.
+  The cloning now happens in the etc/entrypoint.sh script.
 
 v0.3.8
 ------